refactor(decomp): hoist SearchParams + ScanStrategy clusters to crates (gaps 1+2)
Root-crate decomposition: close the first 2 of 6 remaining type-hoist
blockers before src/storage/traits can move to a crate.

Gap 1 — SearchParams (UnifiedSearchParams) cluster -> proximadb-search-types:
  Move the cohesive search-params cluster (UnifiedSearchParams + the
  SearchParams alias + its leaf-type deps: ProgressiveRecalls, SearchMode,
  SearchEffort, VectorFreshnessMode, HybridSearchMode, BlockPruneConfig,
  FilterOptimizationHints) to crates/foundation/proximadb-search-types.
  Every member depends only on foundation types (DistanceMetric,
  FilterExpression, UnifiedQuantizationLevel, BlockPruneMode) + std/serde.
  Root crate::core::search re-exports every type so all callers resolve
  unchanged. The conversion-function modules (protocol_conversions,
  filter_extraction) stay in the root (proto-heavy, not data types).
  Added dep: proximadb-quantization-model (foundation->foundation, allowed).

Gap 2 — ScanStrategy/ScanIterator cluster -> proximadb-storage-ports:
  Move ScanStrategy, ScanIterator, ScanStatistics, ScanCostEstimate to
  crates/storage/proximadb-storage-ports/src/scan_strategy.rs (alongside
  ScanCapabilities hoisted by #910). Pure-data: deps are only FilterExpression
  + ProximaRecord + std. UnifiedScanEngine trait (names proto Collection)
  stays in the root. Added deps: proximadb-filter-expression, proximadb-records.

Fix-ups required by the hoist:
  - src/index/axis/contract.rs: From<IndexSearchEffort> for SearchEffort hit
    the orphan rule (neither type local). Replaced with a free fn
    index_effort_to_search + updated the single .map(Into::into) call site.
  - BlockPruneConfig::for_testing(): un-gated from #[cfg(test)] to
    #[doc(hidden)] pub fn so downstream #[cfg(test)] code can call it
    (the crate's own cfg(test) does not propagate to dependents).
  - FilterOptimizationHints (query_optimizer.rs): replaced def with a
    re-export shim from search-types.

Verification: cargo check --lib --bins green; cargo check --tests -p proximadb
green; nextest -p proximadb-search-types -p proximadb-storage-ports 52/52 pass;
cargo clippy --lib --bins exit 0; fmt --check clean; layering/up-edges gate at
197 (unchanged). 563 root search-related lib tests pass.
